+++ src/Functions/FunctionDateOrDateTimeToSomething.h
@@ -80,7 +80,7 @@ public:
 
     bool hasInformationAboutPreimage() const override { return Transform::hasPreimage(); }
 
-    RangeOrNull getPreimage(const IDataType & type, const Field & point) const override
+    OptionalFieldInterval getPreimage(const IDataType & type, const Field & point) const override
     {
         if constexpr (Transform::hasPreimage())
             return Transform::getPreimage(type, point);
