+++ src/Functions/IFunctionAdaptors.h
@@ -97,7 +97,7 @@ public:
         return function->getMonotonicityForRange(type, left, right);
     }
 
-    RangeOrNull getPreimage(const IDataType & type, const Field & point) const override
+    OptionalFieldInterval getPreimage(const IDataType & type, const Field & point) const override
     {
         return function->getPreimage(type, point);
     }
